refactor: categorize ctl / query loading stuff into their own folders
