+++ src/cli/verifier/CMakeLists.txt
@@ -1,6 +1,6 @@
 project(verifier VERSION 1.0.0)
 find_package(argvparse REQUIRED)
-add_executable(${PROJECT_NAME} main.cpp query_json_loader.cpp)
+add_executable(${PROJECT_NAME} main.cpp query/query_json_loader.cpp)
 target_link_libraries(${PROJECT_NAME} aaltitoad ctl argvparse)
 if(TARGET default_plugins)
     add_dependencies(${PROJECT_NAME} default_plugins)
