+++ src/cli/verifier/main.cpp
@@ -7,7 +7,7 @@
 #include <ntta/interesting_tocker.h>
 #include "cli_options.h"
 #include "../cli_common.h"
-#include "query_json_loader.h"
+#include "query/query_json_loader.h"
 
 auto load_plugins(std::map<std::string, argument_t>& cli_arguments) -> plugin_map_t;
