+++ src/cli/verifier/query/query_json_loader.h
