+++ src/verification/forward_reachability.cpp
@@ -1,5 +1,5 @@
 #include "forward_reachability.h"
-#include "ctl_sat.h"
+#include "verification/ctl/ctl_sat.h"
 
 namespace aaltitoad {
     // TODO: <feature request> todo: add a new issue on github
